# api/openapi.yaml
openapi: "3.1.0"
info:
  title: FamilyRoots API
  version: "1.0.0"
  description: |
    ## FamilyRoots REST API — v1

    A multi-tenant genealogy platform API following REST conventions.

    ### Versioning
    URL-path versioning: all v1 endpoints are prefixed `/api/v1`.
    Breaking changes increment the major version (`/api/v2`).
    Minor/patch changes are backward-compatible and do not change the URL.

    ### Authentication
    Two tokens are issued on login:
    - **Access token** — short-lived JWT (15 min), passed as `Authorization: Bearer <token>`
    - **Refresh token** — long-lived (30 days), stored as an `HttpOnly SameSite=Strict` cookie

    ### Multi-tenancy
    Every request is scoped to a tenant derived from the JWT `tenant_id` claim.
    Tenant is never supplied by the caller — it is always resolved server-side from the token.

    ### Rate limiting
    All endpoints return rate-limit headers:
    - `X-RateLimit-Limit` — requests allowed in the current window
    - `X-RateLimit-Remaining` — requests remaining in the current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    - `Retry-After` — seconds to wait (only present on 429 responses)

    Default limits by plan:
    | Plan        | Requests/min | Burst |
    |-------------|-------------|-------|
    | FREE        | 60          | 10    |
    | BASIC       | 300         | 30    |
    | PREMIUM     | 1000        | 100   |
    | PROFESSIONAL| 5000        | 500   |

...
    ```
...
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued on login. Expires in 15 minutes.
        Payload includes: sub (user_id), tenant_id, email, roles, plan.

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        HttpOnly SameSite=Strict cookie holding the refresh token.
        Used only on POST /auth/refresh and POST /auth/logout.

  # ──────────────────────────────────────────────────────────
  # COMMON PARAMETERS
  # ──────────────────────────────────────────────────────────
  parameters:
    treeId:
      name: treeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Family tree identifier
    personId:
      name: personId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Person identifier
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Maximum number of items to return
    cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response

  # ──────────────────────────────────────────────────────────
  # COMMON HEADERS
  # ──────────────────────────────────────────────────────────
...
  schemas:
...
    PersonSex:
      type: string
      enum: [MALE, FEMALE, OTHER, UNKNOWN]
    PrivacyLevel:
      type: string
      enum: [PUBLIC, FAMILY, PRIVATE]
    NameType:
      type: string
      enum: [BIRTH, MARRIED, ALSO_KNOWN_AS, NICKNAME, RELIGIOUS, ADOPTED, OTHER]
...
    ParentageType:
      type: string
      enum: [BIOLOGICAL, ADOPTIVE, STEP, FOSTER, UNKNOWN]
...
    Error:
      type: object
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri
          example: "https://familyroots.app/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "The request body contains invalid fields."
        instance:
          type: string
          example: "/api/v1/trees/abc/persons"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        requestId:
          type: string
          example: "req_01hzab3k7x"
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          example: "sex"
        code:
          type: string
          example: "INVALID_ENUM"
        message:
          type: string
          example: "Must be one of: MALE, FEMALE, OTHER, UNKNOWN"

    # ── Pagination ──────────────────────────────────────────
    Pagination:
      type: object
      required: [hasMore]
      properties:
        cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page (null if no more pages)
        hasMore:
          type: boolean
        totalCount:
          type: integer
          description: Total matching records (expensive; only returned when requested)

    # ── Auth / Tokens ───────────────────────────────────────
    TokenPair:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          description: Short-lived JWT (15 min)
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          example: 900
          description: Access token TTL in seconds
        user:
          $ref: "#/components/schemas/UserProfile"

    # ── User ────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, email, displayName, isVerified, plan, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          maxLength: 100
        avatarUrl:
          type: string
          format: uri
          nullable: true
        locale:
          type: string
          example: "en"
        timezone:
          type: string
          example: "America/New_York"
        isVerified:
          type: boolean
        plan:
          type: string
          enum: [FREE, BASIC, PREMIUM, FAMILY, PROFESSIONAL]
        createdAt:
          type: string
          format: date-time
...
    PersonName:
      type: object
      required: [id, nameType, isPrimary]
      properties:
        id:
          type: string
          format: uuid
        nameType:
          $ref: "#/components/schemas/NameType"
        givenName:
          type: string
          nullable: true
        surname:
          type: string
          nullable: true
        prefix:
          type: string
          nullable: true
        suffix:
          type: string
          nullable: true
        nickname:
          type: string
          nullable: true
        fullNameText:
          type: string
          nullable: true
          description: Free-form override for non-Western names
        isPrimary:
          type: boolean
        validFrom:
          type: string
          format: date
          nullable: true
        validTo:
          type: string
          format: date
          nullable: true
    Person:
      type: object
      required: [id, treeId, sex, privacyLevel, isLiving, isDeceased, names, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        treeId:
          type: string
          format: uuid
        sex:
          $ref: "#/components/schemas/PersonSex"
        privacyLevel:
          $ref: "#/components/schemas/PrivacyLevel"
        isLiving:
          type: boolean
        isDeceased:
          type: boolean
        displayGivenName:
          type: string
          nullable: true
        displaySurname:
          type: string
          nullable: true
        names:
          type: array
          items:
            $ref: "#/components/schemas/PersonName"
        customFields:
          type: object
          additionalProperties: true
        gedcomXref:
          type: string
          nullable: true
        version:
          type: integer
          description: Optimistic lock version — must be echoed on PATCH requests
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PersonSummary:
      type: object
      description: Lightweight representation used in list responses and tree views
      required: [id, sex, isLiving, isDeceased]
      properties:
        id:
          type: string
          format: uuid
        sex:
          $ref: "#/components/schemas/PersonSex"
        isLiving:
          type: boolean
        isDeceased:
          type: boolean
        displayGivenName:
          type: string
          nullable: true
        displaySurname:
          type: string
          nullable: true
        privacyLevel:
          $ref: "#/components/schemas/PrivacyLevel"
        birthYear:
          type: integer
          nullable: true
          description: Derived from BIRTH event; null if unknown
        deathYear:
          type: integer
          nullable: true
    CreatePersonNameInput:
      type: object
      required: [nameType, isPrimary]
      properties:
        nameType:
          $ref: "#/components/schemas/NameType"
        givenName:
          type: string
          maxLength: 200
          nullable: true
        surname:
          type: string
          maxLength: 200
          nullable: true
        prefix:
          type: string
          maxLength: 50
          nullable: true
        suffix:
          type: string
          maxLength: 50
          nullable: true
        nickname:
          type: string
          maxLength: 200
          nullable: true
        fullNameText:
          type: string
          maxLength: 500
          nullable: true
        isPrimary:
          type: boolean
          default: false
        validFrom:
          type: string
          format: date
          nullable: true
        validTo:
          type: string
          format: date
          nullable: true
    CreatePersonRequest:
      type: object
      required: [names]
      properties:
        sex:
          $ref: "#/components/schemas/PersonSex"
          default: UNKNOWN
        privacyLevel:
          $ref: "#/components/schemas/PrivacyLevel"
          default: FAMILY
        isLiving:
          type: boolean
          default: true
        isDeceased:
          type: boolean
          default: false
        names:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreatePersonNameInput"
          description: At least one name required; exactly one must have isPrimary=true
        customFields:
          type: object
          additionalProperties: true
    UpdatePersonRequest:
      type: object
      required: [version]
      minProperties: 2
      properties:
        version:
          type: integer
          description: Must match current server version (optimistic locking)
        sex:
          $ref: "#/components/schemas/PersonSex"
        privacyLevel:
          $ref: "#/components/schemas/PrivacyLevel"
        isLiving:
          type: boolean
        isDeceased:
          type: boolean
        customFields:
          type: object
          additionalProperties: true
    AncestorNode:
      type: object
      required: [person, depth, familyGroupId]
      properties:
        person:
          $ref: "#/components/schemas/PersonSummary"
        depth:
          type: integer
          description: "1 = parent, 2 = grandparent, etc."
        familyGroupId:
          type: string
          format: uuid
        parentageType:
          $ref: "#/components/schemas/ParentageType"

    # ── Family Group ─────────────────────────────────────────
...
  responses:
    400BadRequest:
      description: Bad Request — malformed JSON or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    401Unauthorized:
      description: Unauthorized — missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    403Forbidden:
      description: Forbidden — valid token but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    422UnprocessableEntity:
      description: Unprocessable Entity — validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# TAGS
# ──────────────────────────────────────────────────────────────
...
paths:
...
  /auth/oauth/{provider}:
    get:
      tags: [Authentication]
      operationId: oauthRedirect
      summary: Redirect to OAuth provider authorisation page
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, facebook, apple]
        - name: redirectUri
          in: query
          required: true
          schema:
            type: string
            format: uri
      responses:
        "302":
          description: Redirect to provider
        "400":
          $ref: "#/components/responses/400BadRequest"
  /auth/oauth/{provider}/callback:
    get:
      tags: [Authentication]
      operationId: oauthCallback
      summary: Handle OAuth provider callback
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, facebook, apple]
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OAuth login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "409":
          description: Email already registered with a different provider

  # ============================================================
  # USERS
  # ============================================================
...
  /trees/{treeId}/persons:
    parameters:
      - $ref: "#/components/parameters/treeId"
    get:
      tags: [Persons]
      operationId: listPersons
      summary: List persons in a tree
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
        - name: surname
          in: query
          schema:
            type: string
          description: Filter by surname (prefix match)
        - name: sex
          in: query
          schema:
            $ref: "#/components/schemas/PersonSex"
        - name: isLiving
          in: query
          schema:
            type: boolean
        - name: orderBy
          in: query
          schema:
            type: string
            enum: [surname, givenName, createdAt, updatedAt]
            default: surname
      responses:
        "200":
          description: Paginated list of persons
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PersonSummary"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

    post:
      tags: [Persons]
      operationId: createPerson
      summary: Create a new person in a tree
      description: Requires CONTRIBUTOR, EDITOR, or ADMIN role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonRequest"
      responses:
        "201":
          description: Person created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "422":
          $ref: "#/components/responses/422UnprocessableEntity"
  /trees/{treeId}/persons/{personId}:
    parameters:
      - $ref: "#/components/parameters/treeId"
      - $ref: "#/components/parameters/personId"
    get:
      tags: [Persons]
      operationId: getPerson
      summary: Get a person by ID
      parameters:
        - name: include
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [names, events, media, sources, familyGroups]
          style: form
          explode: false
          description: Comma-separated list of related resources to embed in the response
      responses:
        "200":
          description: Person detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

    patch:
      tags: [Persons]
      operationId: updatePerson
      summary: Update a person's attributes
      description: >
        Requires CONTRIBUTOR, EDITOR, or ADMIN role.
        The `version` field must match the current server version (optimistic locking).
        On conflict, the server returns 409 with the current version.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePersonRequest"
      responses:
        "200":
          description: Updated person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
        "409":
          description: Version conflict — another edit was saved concurrently
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    properties:
                      currentVersion:
                        type: integer
                        description: The version currently on the server

    delete:
      tags: [Persons]
      operationId: deletePerson
      summary: Soft-delete a person
      description: >
        Sets is_deleted=true. Person remains in version_history.
        Requires EDITOR or ADMIN role.
      responses:
        "204":
          description: Person soft-deleted
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
...
  /trees/{treeId}/persons/{personId}/ancestors:
    parameters:
      - $ref: "#/components/parameters/treeId"
      - $ref: "#/components/parameters/personId"
    get:
      tags: [Persons]
      operationId: getAncestors
      summary: Get all ancestors of a person (unlimited generations)
      parameters:
        - name: maxDepth
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: parentageTypes
          in: query
          schema:
            type: array
            items:
              $ref: "#/components/schemas/ParentageType"
          style: form
          explode: false
          description: "Filter by parentage type (default: all types)"
      responses:
        "200":
          description: Ancestor list ordered by depth
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AncestorNode"
                  subject:
                    $ref: "#/components/schemas/PersonSummary"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
  /trees/{treeId}/persons/{personId}/descendants:
    parameters:
      - $ref: "#/components/parameters/treeId"
      - $ref: "#/components/parameters/personId"
    get:
      tags: [Persons]
      operationId: getDescendants
      summary: Get all descendants of a person (unlimited generations)
      parameters:
        - name: maxDepth
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: parentageTypes
          in: query
          schema:
            type: array
            items:
              $ref: "#/components/schemas/ParentageType"
          style: form
          explode: false
      responses:
        "200":
          description: Descendant list ordered by depth
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AncestorNode"
                  subject:
                    $ref: "#/components/schemas/PersonSummary"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
...